+++ ens-20171110/core/Models/DescribeNetworkAttributeResponseBody.cs
@@ -232,6 +232,16 @@ namespace AlibabaCloud.SDK.Ens20171110.Models
         [Validation(Required=false)]
         public string RouterTableId { get; set; }
 
+        [NameInMap("SecondaryCidrBlocks")]
+        [Validation(Required=false)]
+        public DescribeNetworkAttributeResponseBodySecondaryCidrBlocks SecondaryCidrBlocks { get; set; }
+        public class DescribeNetworkAttributeResponseBodySecondaryCidrBlocks : TeaModel {
+            [NameInMap("SecondaryCidrBlock")]
+            [Validation(Required=false)]
+            public List<string> SecondaryCidrBlock { get; set; }
+
+        }
+
         /// <summary>
         /// <para>The status of the network. Valid values:</para>
         /// <list type="bullet">