+++ ens-20171110/core/Properties/AssemblyInfo.cs
@@ -20,7 +20,7 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("3a4f2ed0-8f32-11f0-b8ad-7b3047cd768b")]
+[assembly: Guid("005fcbc0-9320-11f0-8283-217d839031e0")]
 
 // Version information for an assembly consists of the following four values:
 //
@@ -29,5 +29,5 @@ using System.Runtime.InteropServices;
 //      Build Number
 //      Revision
 //
-[assembly: AssemblyVersion("8.2.0.0")]
-[assembly: AssemblyFileVersion("8.2.1.0")]
+[assembly: AssemblyVersion("8.2.1.0")]
+[assembly: AssemblyFileVersion("8.2.2.0")]